backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
pydantic[email]==2.9.2
celery[redis]==5.4.0
flower==2.0.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
structlog==24.4.0
python-multipart==0.0.12
solders==0.21.0
base58==2.1.1
psycopg2-binary==2.9.10
aiofiles==24.1.0
